test(acp): isolate configured binding regression coverage
Keep the real manager/lifecycle regression in a focused test file and narrow
the adapter case union explicitly. Runtime behavior is unchanged.
